First massive commit towards 3.0
- C99-based
- bool type extensively used
- significantly reduced lexical scoping of variables
- implemented load of non-seekable files (e.g., named pipes)
- no limits on line lengths or number of lines (both are now 64-bit)
- standard char type for buffer contents (no more unsigned)
- only -Wno-parentheses -Wno-switch warning disabling

+++ src/syntax.h
@@ -100,8 +100,8 @@ struct high_syntax *load_syntax PARAMS((unsigned char *name));
 
 /* Parse a lines.  Returns new state. */
 
-extern int *attr_buf;
-extern int attr_len;
+extern uint32_t *attr_buf;
+extern int64_t attr_len;
 HIGHLIGHT_STATE parse PARAMS((struct high_syntax *syntax, line_desc *ld, HIGHLIGHT_STATE h_state, int utf8));
 
 #define clear_state(s) (((s)->saved_s[0] = 0), ((s)->state = 0), ((s)->stack = 0))
